+++ packages/bpk-tokens/tokens/ios/base.react.native.es6.js
@@ -17,6 +17,9 @@
  * limitations under the License.
  * 
  */
+export const borderSizeSm = 1;
+export const borderSizeLg = 2;
+export const borderSizeXl = 3;
 export const colorWhite = "rgb(255, 255, 255)";
 export const colorBlue50 = "rgb(225, 244, 248)";
 export const colorBlue100 = "rgb(203, 238, 245)";
@@ -116,6 +119,11 @@ export const textXxlFontSize = 34;
 export const textXxlFontWeight = "700";
 export const textXxlLineHeight = 41;
 export const textEmphasizedFontWeight = "600";
+export const borders = {
+borderSizeSm,
+borderSizeLg,
+borderSizeXl,
+};
 export const colors = {
 colorWhite,
 colorBlue50,